requirements.txt
pyqt
anytree